[Fix] decode URL encoded string in config path
Some packages has config paths which is url encoded (e.g. %2B for +)
https://github.com/open-mmlab/mmdetection/blob/db85fd12afce2fe89d1c7b870874c02a90018a16/configs/gn%2Bws/metafile.yml#L17

Currently downloading those models trrigers "is not found" error
because mim does not support url encoded path.

While I am not sure if this should be fixed in mim or in metafile.yml in
the package, I firstly attempt to fix in mim package.